+++ changelog.md
@@ -1,6 +1,7 @@
 # UNRELEASED
 
 ### Features
+
 * [3672](https://github.com/zeta-chain/node/pull/3672) - zetaclient: cache tss signatures for performance.
 
 # CHANGELOG