+++ changelog.md
@@ -1,3 +1,9 @@
+# UNRELEASED
+
+### Features
+
+* [3672](https://github.com/zeta-chain/node/pull/3672) - zetaclient: cache tss signatures for performance.
+
 # CHANGELOG
 
 ## v29.0.0